Car Rental system using C++ with  file handling.
The concepts used here are 
-Object Oriented Programming(OOP)
-Inheritance (Multilevel, Multiple, Hierarchical)
-File Handling (CSV)
-Functions(Call by value, Call by reference)

While this code may not be perfect, It covers all functionalities of the given problem statement and I had to finish it within 6 hours. 
You Can Modify the names inside the C++ code and use it if you like. Make sure to change the names in 'data.txt' as well.